# app/api_contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_clone_voices_upload_api_clone_voices_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "ref_text": {
            "default": "",
            "title": "Ref Text",
            "type": "string"
          },
          "rights_basis": {
            "default": "",
            "title": "Rights Basis",
            "type": "string"
          },
          "rights_confirmed": {
            "default": false,
            "title": "Rights Confirmed",
            "type": "boolean"
          },
          "source_title": {
            "default": "",
            "title": "Source Title",
            "type": "string"
          },
          "source_url": {
            "default": "",
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_clone_voices_upload_api_clone_voices_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoiceDesignPreviewRequest": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "sample_text": {
            "title": "Sample Text",
            "type": "string"
          }
        },
        "required": [
          "description",
          "sample_text"
        ],
        "title": "VoiceDesignPreviewRequest",
        "type": "object"
      },
      "VoiceDesignSaveRequest": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preview_file": {
            "title": "Preview File",
            "type": "string"
          },
          "sample_text": {
            "title": "Sample Text",
            "type": "string"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          }
        },
        "required": [
          "name",
          "description",
          "sample_text",
          "preview_file"
        ],
        "title": "VoiceDesignSaveRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Alexandria Audiobook",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/clone_voices/upload": {
      "post": {
        "description": "Import a reference clip for voice cloning.\n\nThe clip is decoded, normalised to what the engine reads (24 kHz mono\nPCM16 WAV), measured, and refused if it is not one clean 3-30 s sentence\n(see voice_reference_import). The exact transcript and a rights\nconfirmation are required; source title/URL/basis are recorded when given.\nEvery import is a new immutable id - a changed reference is a new voice.",
        "operationId": "clone_voices_upload_api_clone_voices_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_clone_voices_upload_api_clone_voices_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clone Voices Upload"
      }
    },
    "/api/clone_voices/{voice_id}": {
      "delete": {
        "description": "Delete an uploaded clone voice.",
        "operationId": "clone_voices_delete_api_clone_voices__voice_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "voice_id",
            "required": true,
            "schema": {
              "title": "Voice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clone Voices Delete"
      }
    },
...
    "/api/voice_design/preview": {
      "post": {
        "description": "Generate a preview voice from a text description.",
        "operationId": "voice_design_preview_api_voice_design_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceDesignPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Voice Design Preview"
      }
    },
    "/api/voice_design/save": {
      "post": {
        "description": "Save a preview voice as a permanent designed voice.",
        "operationId": "voice_design_save_api_voice_design_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceDesignSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Voice Design Save"
      }
    },
    "/api/voice_design/{voice_id}": {
      "delete": {
        "description": "Delete a saved designed voice.",
        "operationId": "voice_design_delete_api_voice_design__voice_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "voice_id",
            "required": true,
            "schema": {
              "title": "Voice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Voice Design Delete"
      }
    },
...
  }
}